+++ CHANGELOG.md
@@ -1,10 +1,16 @@
 # Changelog
 
 ## [Unreleased]
+- Added support for Keepsakes.
+- Add check for paths to remove `_PLUGIN.guid` if passed in the file paths.
+- Added missed flavour text for SJSON.
+- Defs.
+- Update ReadME.
+
 
 ## [1.0.2] - 2025-11-09
 
-- No longer need to call `gods.Initialize()`, just pass the \_PLUGIN.guid into SJSON creation.
+- No longer need to call `gods.Initialize()`, just pass the `_PLUGIN.guid` into SJSON creation.
 
 ## [1.0.1] - 2025-11-07